requirements.txt
python-binance==1.0.17
requests==2.28.1
